backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.2",
    "uvicorn[standard]>=0.40.0",
]
